using glfw callbacks; some playing around with values

+++ src/renderable.h
@@ -92,7 +92,7 @@ public:
         glVertexAttribPointer(shader["vVertex"], 3, GL_FLOAT, GL_FALSE,0,0);
 
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, vboIndicesID);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, totalIndices * sizeof(GLuint), 0, GL_STATIC_DRAW);
+        glBufferData(GL_ELEMENT_ARRAY_BUFFER, totalIndices * sizeof(GLshort), 0, GL_STATIC_DRAW);
 
         GLuint* pIBuffer = static_cast<GLuint*>(glMapBuffer(GL_ELEMENT_ARRAY_BUFFER,
                                                             GL_WRITE_ONLY));
